week 9; going to set up database on production

+++ bookmark/database/migrations/2020_03_30_162411_create_books_table.php
@@ -28,7 +28,7 @@ class CreateBooksTable extends Migration
             $table->string('author');
 
             // published_year TINYINT
-            $table->tinyInteger('published_year');
+            $table->smallInteger('published_year');
 
             // cover_url VARCHAR
             $table->string('cover_url');
